# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Human resource management (HRM) is the practice of recruiting, hiring, deploying and managing an organization's employees. HRM is often referred to simply as human resources (HR). A company or organization's HR department is usually responsible for creating, putting into effect and overseeing policies governing workers and the relationship of the organization with its employees. ",
    "title": "HR Management"
  },
  "schemas": ["http"],
  "host": "localhost:3000",
  "basePath": "/api/",
  "tags": [
    {
      "name": "Employee",
      "description": "Employee management"
    },
    { "name": "store", "description": "Access to Petstore orders" },
    {
      "name": "Team",
      "description": "Team management"
    }
  ],
  "schemes": ["http"],
  "paths": {
    "/employees/": {
      "get": {
        "tags": ["Employee"],
        "summary": "get all employees",
        "description": "",
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": { "$ref": "#/definitions/ApiResponse" }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/employees/login": {
      "post": {
        "tags": ["Employee"],
        "summary": "Logs user into the system",
        "description": "",
        "operationId": "loginUser",
        "produces": ["application/json", "application/xml"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Login",
            "required": true,
            "schema": { "$ref": "#/definitions/Login" }
          }
        ],
        "responses": {
          "200": {
            "token": "a78s6d76a8wf5a6w8f56as75f67awf76as98f79w6ef7aw6e87f6s7f698d",
            "description": "successful operation",
            "headers": {
              "X-Expires-After": {
                "type": "string",
                "format": "date-time",
                "description": "date in UTC when token expires"
              }
            },
            "userId": "123eqw7as78df6a",
            "role": "ADMIN",
            "schema": { "type": "string" }
          },
          "400": { "description": "Invalid username/password supplied" }
        }
      }
    }
  },
  "securityDefinitions": {
    "bearerAuth": {
      "name": "Authorization",
      "in": "header",
      "type": "apiKey",
      "description": "JWT Authorization header"
    }
  },
  "definitions": {
    "Team": {
      "type": "object",
      "properties": {
        "_id": { "type": "string" },
        "name": { "type": "string" },
        "leader": { "$ref": "#/definitions/Employee" },
        "createdBy": { "$ref": "#/definitions/Employee" },
        "employees": {
          "type": "array",
          "xml": { "wrapped": true },
          "items": {
            "xml": { "name": "employee" },
            "$ref": "#/definitions/Employee"
          }
        },
        "isDeleted": { "type": "boolean" },
        "createdDate": { "type": "date" },
        "modifiedDate": { "type": "date" }
      },
      "xml": { "name": "Team" }
    },
    "Employee": {
      "type": "object",
      "properties": {
        "_id": { "type": "string" },
        "email": { "type": "string" },
        "category": { "$ref": "#/definitions/Category" },
        "name": { "type": "string" },
        "dob": { "type": "date" },
        "phoneNumber": { "type": "number" },
        "": {
          "type": "array",
          "xml": { "wrapped": true },
          "items": { "type": "string", "xml": { "name": "photoUrl" } }
        },
        "currentTeams": {
          "type": "array",
          "xml": { "wrapped": true },
          "items": { "xml": { "name": "team" }, "$ref": "#/definitions/Team" }
        },
        "status": {
          "type": "string",
          "description": "employee status",
          "enum": ["AVAILABLE", "UNAVAILABLE"]
        },
        "role": {
          "type": "string",
          "description": "employee's role",
          "enum": ["ADMIN", "MEMBER", "LEADER"]
        },
        "isDeleted": { "type": "boolean" },
        "createdDate": { "type": "date" },
        "modifiedDate": { "type": "date" }
      },
      "xml": { "name": "Employee" }
    },
    "Login": {
      "type": "object",
      "properties": {
        "email": { "type": "string" },
        "password": { "type": "string" }
      },
      "xml": { "name": "Employee" }
    }
  }
}